refactor: move enterprise apis to insomnia-api
