+++ packages/insomnia-api/src/index.ts
@@ -1,5 +1,6 @@
 export * from './user';
 export * from './vault';
+export * from './enterprise';
 export * from './trial';
 
 export { configureFetch, type FetchConfig } from './fetch';
